Issue #985 Move types into its own file
